% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip.R
\name{antecedent_precip}
\alias{antecedent_precip}
\title{Compute antecedent precipitation}
\usage{
antecedent_precip(x, period = 48, delay = 0, fun = sum,
  value.name = "Precip", datetime.name = "Datetime")
}
\arguments{
\item{x}{Dataframe of hourly precipitation values}

\item{period}{Duration of antecedent period as hours (default=48)}

\item{delay}{Shift in the antecedent period as hours (default=0)}

\item{fun}{Function applied to each period (e.g. sum, max)}

\item{value.name}{Name of precipitation column}

\item{datetime.name}{Name of date/time column}
}
\value{
numeric vector of antecedent precipitation for given period, delay, and fun
}
\description{
Computes antecedent precipitation as a rolling sum. Time series
must be continuous at hourly intervals
}
\examples{
filepath <- system.file("extdata", "LoganPrecip.xlsx", package = "myrwaR")
pcp <- load_precip_from_xls(filepath, as.type='dataframe')
pcp$Precip.48 <- antecedent_precip(pcp, period=48, delay=0,
                                   datetime.name="Datetime",
                                   value.name="Precip")
}

