% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/events.R
\name{assign_precip_events}
\alias{assign_precip_events}
\title{Delineates storm events based on hourly precipitation}
\usage{
assign_precip_events(x, datetime.name = "Datetime", value.name = "Precip",
  interevent.period = 8, threshold.total = 0.1)
}
\arguments{
\item{x}{Dataframe containing hourly precipitation timeseries}

\item{datetime.name}{Name of datetime column}

\item{value.name}{Name of precipitation value column}

\item{interevent.period}{Minimum inter-event period to distinguish separate
events}

\item{threshold.total}{Minimum total precipitation for an individual event}
}
\value{
dataframe or zoo object of hourly precipitation values
}
\description{
Uses a timeseries of hourly precipitation to define storm events and create
summary table
}

