% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_no_empty_strings}
\alias{check_no_empty_strings}
\title{Checks for empty strings}
\usage{
check_no_empty_strings(x)
}
\arguments{
\item{x}{character vector}
}
\value{
boolean
}
\description{
Given character vector, checks if any of the unique values are empty strings. Prints message listing any empty string values.
}
\examples{
check_string_trim(c("a", "b", " "))  # false
check_string_trim(c("a", "b", "  ")) # false
}

