% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{check_string_untrimmed}
\alias{check_string_untrimmed}
\title{Checks for untrimmed strings}
\usage{
check_string_untrimmed(x)
}
\arguments{
\item{x}{character vector}
}
\value{
boolean
}
\description{
Given character vector, checks if any of the unique values have extra spaces
at the start or end (e.g. "a " or " a").
Prints message listing any untrimmed values.
}
\examples{
check_string_trim(c("a", "b", "c ")) # false
}

