% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qaqc.R
\name{compare_database}
\alias{compare_database}
\title{Compare two database versions and summarize differences}
\usage{
compare_database(old_path, new_path)
}
\arguments{
\item{old_path}{Path to older database version}

\item{new_path}{Path to newest database version}
}
\description{
Loads data from two separate databases and prints a summary of which records
have been added or removed. Note this does not check for changes to
individual values, it only looks at which IDs have been added or removed.
}

