% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_list_tables}
\alias{db_list_tables}
\title{List of table names in database}
\usage{
db_list_tables(ch)
}
\arguments{
\item{ch}{Connection handle to database}
}
\value{
A character vector of table names
}
\description{
Retrieves a character vector of table names
}
\examples{
ch <- db_connect("D:/Dropbox/Work/mystic/db/MysticDB_20160208.accdb")
table_names <- db_list_tables(ch)
}

