% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_locations}
\alias{db_locations}
\title{Retrieve locations from database}
\usage{
db_locations(ch)
}
\arguments{
\item{ch}{Connection handle to database}
}
\value{
A dataframe containing the locations in the database
}
\description{
Retrieves the station locations from the database by merging Location,
WaterBody, and LocationType tables
}
\examples{
ch <- db_connect("D:/Dropbox/Work/mystic/db/MysticDB_20160208.accdb")
locations <- db_locations(ch)
}

