% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{db_qaqc_suite}
\alias{db_qaqc_suite}
\title{Run QAQC test suite on database}
\usage{
db_qaqc_suite(ch, print.rows)
}
\arguments{
\item{ch}{Connection handle to database}

\item{print.rows}{If TRUE, prints individual rows that did not pass any given test (default=FALSE)}
}
\value{
nothing
}
\description{
Runs a suite of tests to check for invalid values, relationships, and other issues in the database.
}
\examples{
db_qaqc_suite(ch)
}

