% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_results}
\alias{db_results}
\title{Load Results table merged with Visit table}
\usage{
db_results(ch, ...)
}
\arguments{
\item{ch}{Open connection handle to database}

\item{...}{Additional arguments passed to db_table()}
}
\value{
Data frame of water quality data containing visit information
}
\description{
Loads the water quality data from the Result table merged with Visit table
}
\examples{
ch <- db_connect("D:/Dropbox/Work/mystic/db/MysticDB_20160208.accdb")
wq <- db_results(ch)
}

