% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_table}
\alias{db_table}
\title{Load single table from database}
\usage{
db_table(ch, table_name, ...)
}
\arguments{
\item{ch}{Connection handle to database}

\item{table_name}{Table name}

\item{...}{Additional arguments passed to sqlFetch()}
}
\value{
A dataframe containing the database table
}
\description{
Loads the data from a table in the database
}
\examples{
ch <- db_connect("D:/Dropbox/Work/mystic/db/MysticDB_20160208.accdb")
tbl.Results <- db_table(ch, "Results")
}

