% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_table_fields}
\alias{db_table_fields}
\title{Retrieve table schema from database}
\usage{
db_table_fields(ch, table_name)
}
\arguments{
\item{ch}{Connection handle to database}
}
\value{
A dataframe containing the column names and types for the specified
  table
}
\description{
Retrieves the field names and types from a given database table
}
\examples{
ch <- db_connect("D:/Dropbox/Work/mystic/db/MysticDB_20160208.accdb")
schema_results <- db_table_fields(ch, table_name = "Results")
}

