% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storet.R
\name{get_storet_results}
\alias{get_storet_results}
\title{Fetch MyRWA Results Table from STORET}
\usage{
get_storet_results(verbose = TRUE)
}
\arguments{
\item{verbose}{If TRUE (default), prints information about the download}
}
\value{
data frame of results
}
\description{
Fetches MyRWA locations from STORET using the http://waterqualitydata.us REST API.
Locations are downloaded to a temporary zip file, unzipped, and loaded into an R
data frame.
}
\examples{
storet_results <- get_storet_results()

}

