% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{is.regular_hourly}
\alias{is.regular_hourly}
\title{Checks if vector of hourly timestamps is regular}
\usage{
is.regular_hourly(x)
}
\arguments{
\item{x}{zoo object or vector of datetimes}
}
\value{
boolean
}
\description{
Given a timeseries as a zoo object or vector of datetime values, checks if
the series is regular (i.e. consecutive and complete) at hourly timesteps
}
\examples{
is.regular_hourly(df$Datetime)
}

