% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{load_wq}
\alias{load_wq}
\title{Load water quality data}
\usage{
load_wq(path, projects = NULL, sample_types = "S", exclude_flags = FALSE)
}
\arguments{
\item{path}{Full path to MyRWA Access Database}

\item{projects}{vector of ProjectIDs to keep (default=NULL to return all projects)}

\item{sample_types}{SampleTypeID(s) to keep in dataset (default="S" to
exclude blanks and duplicates, if NULL returns all sample types)}

\item{exclude_flags}{If TRUE, excludes all samples with any flag,
otherwise returns all samples}
}
\value{
A dataframe containing water quality data
}
\description{
Loads water quality Results table merged with Visit and Location tables
}
\examples{
df <- load_wq(path = "D:/Dropbox/Work/mystic/db/MysticDB_20160208.accdb")
# all baseline samples including flagged and qaqc samples
baseline <- load_wq(path = "D:/Dropbox/Work/mystic/db/MysticDB_20160208.accdb", projects="BASE", sample_types=NULL, exclude_flags=FALSE)
}

