% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrwaR.R
\docType{package}
\name{myrwaR}
\alias{myrwaR}
\alias{myrwaR-package}
\title{myrwaR: Tools for the Mystic River Watershed Association}
\description{
The myrwaR package contains sets of functions for loading data from the MyRWA
Access database and analyzing hydrologic and water quality data. The package
also includes \code{rmarkdown} templates for automated report generation.
}
\section{Database Functions}{

\code{\link{db_connect}}: Creates a connection handler for the Access database
}

