% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{precip_event_summary}
\alias{precip_event_summary}
\title{Summarize storm events}
\usage{
precip_event_summary(x, datetime.name = "Datetime", value.name = "Precip")
}
\arguments{
\item{x}{Dataframe containing hourly precipitation and already processed by assign_precip_events()}

\item{datetime.name}{Name of datetime column}

\item{value.name}{Name of precipitation column}
}
\value{
dataframe
}
\description{
Creates a summary table of individual storm events based on hourly precipitation. The
input dataframe must contain the EventID and EventType columns generated
by assign_precip_events().
}

