% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{precip_logan}
\alias{precip_logan}
\title{Hourly Precipitation at Logan Airport, Boston, MA}
\format{A data frame with 17,520 rows and 2 columns:
\describe{
 \item{Datetime}{Datetime as POSIXct with tz="EST"}
 \item{Precip}{Hourly precipitation depth in inches/hour}
}}
\usage{
precip_logan
}
\description{
A dataset containing hourly precipitation at Logan Airport from 2010-2011
}
\examples{
precip_logan
}

