% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qaqc.R
\name{qaqc_table_results}
\alias{qaqc_table_results}
\title{Runs batch of QAQC tests on Results table}
\usage{
qaqc_table_results(ch)
}
\arguments{
\item{ch}{connection handle to database}
}
\value{
nothing
}
\description{
Checks for empty and untrimmed strings
}
\examples{
qaqc_table_results(ch)
}

