% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storet.R
\name{storet_compare_count}
\alias{storet_compare_count}
\title{Compare sample counts between STORET and Database}
\usage{
storet_compare_count(storet, db, name = "VAR")
}
\arguments{
\item{storet}{Vector from STORET}

\item{db}{Vector from Database}

\item{name}{Name of the variable}
}
\value{
nothing
}
\description{
Compares the number of samples per unique value in a single column between STORET and the database.
}

