% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storet.R
\name{storet_results}
\alias{storet_results}
\title{Fetch MyRWA Results Table from STORET}
\usage{
storet_results(verbose = TRUE)
}
\arguments{
\item{verbose}{If TRUE (default), prints information about the download}
}
\value{
data frame of results
}
\description{
Fetches MyRWA locations from STORET using the http://waterqualitydata.us REST API.
Locations are downloaded to a temporary zip file, unzipped, and loaded into an R
data frame.
}

