% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqx.R
\name{wqx_locations}
\alias{wqx_locations}
\title{Convert Locations to WQX Format}
\usage{
wqx_locations(ch, locations)
}
\arguments{
\item{ch}{Open connection handle to database}

\item{locations}{Vector of LocationIDs}
}
\value{
data frame of locations in WQX format
}
\description{
Extracts locations from database and returns data frame in WQX format
}

