% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqx.R
\name{wqx_projects}
\alias{wqx_projects}
\title{Convert Projects to WQX Format}
\usage{
wqx_projects(ch, projects)
}
\arguments{
\item{ch}{Open connection handle to database}

\item{projects}{Vector of ProjectIDs}
}
\value{
data frame of projects in WQX format
}
\description{
Extracts projects from database and returns data frame in WQX format
}

