% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wqx.R
\name{wqx_validate_locations}
\alias{wqx_validate_locations}
\title{Validate Locations Table}
\usage{
wqx_validate_locations(x)
}
\arguments{
\item{x}{Dataframe of locations table}
}
\value{
nothing
}
\description{
Checks for missing columns, missing values, and erroneous values not in
WQX domain. If an error occurs, a warning message is shown.
}

