% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqx.R
\name{wqx_validate_results}
\alias{wqx_validate_results}
\title{Validate Results Table}
\usage{
wqx_validate_results(x)
}
\arguments{
\item{x}{Dataframe of results table}
}
\value{
nothing
}
\description{
Checks for missing columns, missing values, and erroneous values not in
WQX domain. If an error occurs, a warning message is shown.
}

