% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{zoo.regular}
\alias{zoo.regular}
\title{Constructs regular zoo object}
\usage{
zoo.regular(dates, values, by = "hour", fill = NA)
}
\arguments{
\item{dates}{vector of POSIXct values}

\item{values}{vector of numeric values}

\item{by}{target frequency (e.g. "hour", "day", etc), see seq.POSIXt}

\item{fill}{value to fill missing elements}
}
\value{
boolean
}
\description{
Given vector of datetimes and values, constructs regular time series with filling
}
\examples{
zoo.regular(df$Datetime, df$Value, by="hour", fill=NA)
}

