% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/precip.R
\name{antecedent_precip}
\alias{antecedent_precip}
\title{Computes antecedent precipitation}
\usage{
antecedent_precip(x, period = 48, delay = 0, fun = sum,
  value.name = "Precip", datetime.name = "Datetime")
}
\arguments{
\item{x}{data frame of hourly precipitation values}

\item{period}{duration of antecedent period in number of hours}

\item{delay}{shift the antecedent period in number of hours}

\item{fun}{function applied to each period (sum, max)}

\item{value.name}{name of values column}

\item{datetime.name}{name of date/time column}
}
\value{
numeric vector of antecedent precipitation
}
\description{
Computes antecedent precipitation as a rolling sum. Ensures the time series
is regular and continuous
}
\examples{
pcp$Precip.48 <- antecedent_precip(pcp, period=48, delay=0,
                                   datetime.name="Datetime",
                                   value.name="Precip")
}

