% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/precip.R
\name{antecedent_precip}
\alias{antecedent_precip}
\title{Computes antecedent precipitation}
\usage{
antecedent_precip(x, period = 48, precip.name = "Precip",
  datetime.name = "Datetime")
}
\arguments{
\item{x}{data frame of hourly precipitation values}

\item{period}{antecedent period in number of hours}

\item{precip.name}{name of precipitation values column}

\item{datetime.name}{name of date/time column}
}
\value{
numeric vector of antecedent precipitation
}
\description{
Computes antecedent precipitation as a rolling sum. Ensures the time series is regular and continuous
}
\examples{
pcp$Precip.48 <- antecedent_precip(pcp, period=48, precip.name="Precip", datetime.name="Datetime")
}

