% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{check_hourly}
\alias{check_hourly}
\title{Checks if vector of hourly timestamps is regular}
\usage{
check_hourly(x)
}
\arguments{
\item{x}{vector of POSIXct values}
}
\value{
boolean
}
\description{
Given vector of POSIXct values, checks if the series is regular (i.e. consecutive and complete)
}
\examples{
check_hourly(df$Datetime)
}

