% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{db_connect}
\alias{db_connect}
\title{Connect to Access database}
\usage{
db_connect(db_path)
}
\arguments{
\item{db_path}{Path and filename to database}
}
\value{
Connection handle
}
\description{
Opens a connection to an access database
}
\examples{
ch <- db_connect('C://Users//Jeff//Dropbox//Work//mystic//data//MysticDB_20140510.accdb')
}

