% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{db_connect}
\alias{db_connect}
\title{Connect to Access database}
\usage{
db_connect(path)
}
\arguments{
\item{path}{Path to database}
}
\value{
RODBC connection handle
}
\description{
Opens a connection to an access database
}
\examples{
directory <- "C://Users//Jeff//Dropbox//Work//mystic//data//"
ch <- db_connect(file.path(directory, "MysticDB_20140510.accdb"))
}

