% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{db_connect}
\alias{db_connect}
\title{Connect to Access database}
\usage{
db_connect(path)
}
\arguments{
\item{path}{Path to database}
}
\value{
Connection handle
}
\description{
Opens a connection to an access database
}
\examples{
ch <- db_connect('C://Users//Jeff//Dropbox//Work//mystic//data//MysticDB_20140510.accdb')
}

