% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_list_tables}
\alias{db_list_tables}
\title{Retrieve table names in database}
\usage{
db_list_tables(ch)
}
\arguments{
\item{ch}{Connection handle to database}
}
\value{
A character vector of table names
}
\description{
Retrieves a character vector of table names
}
\examples{
directory <- "C://Users//Jeff//Dropbox//Work//mystic//data//"
ch <- db_connect(file.path(directory, "MysticDB_20140510.accdb"))
table_names <- db_list_tables(ch)
}

