% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{db_list_tables}
\alias{db_list_tables}
\title{Retrieve table names in database}
\usage{
db_list_tables(ch)
}
\arguments{
\item{ch}{Connection handle to database}
}
\description{
Retrieves a character vector of table names
}
\examples{
ch <- db_connect('C://Users//Jeff//Dropbox//Work//mystic//data//MysticDB_20140510.accdb')
table_names <- db_list_tables(ch)
}

