% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{db_locations}
\alias{db_locations}
\title{Retrieve locations from database}
\usage{
db_locations(ch)
}
\arguments{
\item{ch}{Connection handle to database}
}
\value{
A dataframe containing the locations in the database
}
\description{
Retrieves the station locations from the database by merging Location,
WaterBody, and LocationType tables
}
\examples{
directory <- "C://Users//Jeff//Dropbox//Work//mystic//data//"
ch <- db_connect(file.path(directory, "MysticDB_20140510.accdb"))
locations <- db_locations(ch)
}

