% Generated by roxygen2 (4.0.2): do not edit by hand
\name{db_locations}
\alias{db_locations}
\title{Retrieve locations from database}
\usage{
db_locations(ch)
}
\arguments{
\item{ch}{Connection handle to database}
}
\description{
Retrieves the station locations from the database by merging Location, WaterBody, and LocationType tables
}
\examples{
ch <- db_connect('C://Users//Jeff//Dropbox//Work//mystic//data//MysticDB_20140510.accdb')
locations <- db_locations(ch)
}

