% Generated by roxygen2 (4.0.2): do not edit by hand
\name{db_results}
\alias{db_results}
\title{Load water quality data}
\usage{
db_results(ch)
}
\arguments{
\item{ch}{Connection handle to database}
}
\description{
Loads the water quality data from the Result table merged with the Visit table
}
\examples{
ch <- db_connect('C://Users//Jeff//Dropbox//Work//mystic//data//MysticDB_20140510.accdb')
wq <- db_results(ch)
}

