% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{db_results}
\alias{db_results}
\title{Load results table}
\usage{
db_results(ch, ...)
}
\arguments{
\item{ch}{Open connection handle to database}

\item{...}{Additional arguments passed to db_table()}
}
\value{
Data frame of water quality data
}
\description{
Loads the water quality data from the Result table merged with Visit table
}
\examples{
directory <- "C://Users//Jeff//Dropbox//Work//mystic//data//"
ch <- db_connect(file.path(directory, "MysticDB_20140510.accdb"))
wq <- db_results(ch)
}

