% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{db_table}
\alias{db_table}
\title{Load table from database}
\usage{
db_table(ch, table_name)
}
\arguments{
\item{ch}{Connection handle to database}
}
\description{
Loads the data from a table in the database
}
\examples{
ch <- db_connect('C://Users//Jeff//Dropbox//Work//mystic//data//MysticDB_20140510.accdb')
tbl.Results <- db_table(ch, 'Results')
}

