% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{db_table_fields}
\alias{db_table_fields}
\title{Retrieve table schema from database}
\usage{
db_table_fields(ch, table_name)
}
\arguments{
\item{ch}{Connection handle to database}
}
\value{
A dataframe containing the column names and types for the specified
  table
}
\description{
Retrieves the field names and types from a given database table
}
\examples{
directory <- "C://Users//Jeff//Dropbox//Work//mystic//data//"
ch <- db_connect(file.path(directory, "MysticDB_20140510.accdb"))
schema_results <- db_table_fields(ch, table_name = "Results")
}

