% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{db_table_fields}
\alias{db_table_fields}
\title{Retrieve table schema from database}
\usage{
db_table_fields(ch, table_name)
}
\arguments{
\item{ch}{Connection handle to database}
}
\description{
Retrieves the field names and types from a given database table
}
\examples{
ch <- db_connect('C://Users//Jeff//Dropbox//Work//mystic//data//MysticDB_20140510.accdb')
schema_results <- db_table_fields(ch, 'Results')
}

