% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip.R
\name{load_precip_from_usgs}
\alias{load_precip_from_usgs}
\title{Load precipitation from USGS Web Services}
\usage{
load_precip_from_usgs(start_date, end_date = NULL, station_id = "01104683",
  tz = "EST", as.type = c("dataframe", "zoo"), datetime.name = "Datetime",
  value.name = "Precip")
}
\arguments{
\item{start_date}{Start date of timeseries as string}

\item{end_date}{End date of timeseries as string}

\item{station_id}{USGS station ID (default=01104683 for Muddy River)}

\item{tz}{Timezone assigned to resulting dataframe (default="EST")}

\item{as.type}{Return type as 'dataframe' (default) or 'zoo'}
}
\value{
dataframe or zoo object of hourly precipitation timeseries in
  inches/hour
}
\description{
Fetches and returns precipitation timeseries from USGS web services.
Timeseries is automatically aggregated to hourly timesteps.
}
\examples{
pcp_usgs <- load_precip_from_usgs(start_date="2015-05-01",
                                  end_date="2015-05-10")
}

