% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/precip.R
\name{load_precip_from_usgs}
\alias{load_precip_from_usgs}
\title{Load precipitation from USGS Web Services}
\usage{
load_precip_from_usgs(startDate, endDate, siteNumber = "01104683",
  tz = "EST", as.type = c("dataframe", "zoo"))
}
\arguments{
\item{startDate}{Start date of timeseries as string}

\item{endDate}{End date of timeseries as string}

\item{siteNumber}{USGS station ID (default=01104683 for Muddy River)}

\item{tz}{Timezone assigned to resulting dataframe (default="EST")}

\item{as.type}{Return type as 'dataframe' (default) or 'zoo'}
}
\value{
dataframe or zoo object of hourly timeseries
}
\description{
Fetches and returns precipitation timeseries from USGS web services.
Timeseries is automatically aggregated to hourly timesteps.
}
\examples{
pcp_usgs <- load_precip_from_usgs(startDate="2015-05-01",
                                  endDate="2015-05-10")
}

