% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{load_precip_from_xls}
\alias{load_precip_from_xls}
\title{Load precipitation from LoganPrecip.xlsx}
\usage{
load_precip_from_xls(path, as.type = c("dataframe", "zoo"))
}
\arguments{
\item{path}{path and filename to precipitation spreadsheet (LoganPrecip.xlsx)}

\item{as.type}{return type as 'dataframe' (default) or 'zoo'}
}
\value{
dataframe or zoo object of hourly precipitation values
}
\description{
Loads the hourly precipitation data from LoganPrecip.xlsx. Returns zoo object or data frame with columns (Datetime, Precip)
}
\examples{
pcp.df <- load_precip_from_xls(file.path("..","..","_share_db","Processed","Precip","LoganPrecip.xlsx"), as.type='dataframe')
pcp.zoo <- load_precip_from_xls(file.path("..","..","_share_db","Processed","Precip","LoganPrecip.xlsx"), as.type='zoo')
}

