% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip.R
\name{load_precip_from_xls}
\alias{load_precip_from_xls}
\title{Load precipitation dataset from Excel File}
\usage{
load_precip_from_xls(path, sheet.name = "Processed precipitation",
  tz = "EST", datetime.name = "Datetime", value.name = "Precip",
  as.type = c("dataframe", "zoo"))
}
\arguments{
\item{path}{Full path and filename to precipitation spreadsheet (LoganPrecip.xlsx)}

\item{sheet.name}{Name of workbook sheet to load data from (default="Processed precipitation")}

\item{tz}{Timezone of date/time stamps (default="EST")}

\item{datetime.name}{Name of datetime column (default="Datetime")}

\item{value.name}{Name of precipitation column (default="Precip")}

\item{as.type}{Return type as 'dataframe' (default) or 'zoo'}
}
\value{
dataframe or zoo object of hourly precipitation values
}
\description{
Loads the hourly precipitation data from an Excel spreadsheet
(specifically: LoganPrecip.xlsx).
}
\examples{
filepath <- system.file("extdata", "LoganPrecip.xlsx", package = "myrwaR")
pcp.df <- load_precip_from_xls(filepath, as.type='dataframe')
pcp.zoo <- load_precip_from_xls(filepath, as.type='zoo')
}

