% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{load_wq}
\alias{load_wq}
\title{Load water quality data}
\usage{
load_wq(path, projects = NULL, sample_types = "S", exclude_flags = FALSE)
}
\arguments{
\item{path}{Path to MyRWA Access Database}

\item{projects}{ProjectID(s) to keep in dataset (default=NULL for all projects)}

\item{sample_types}{SampleTypeID(s) to keep in dataset (default='S' to exclude blanks and duplicates, if NULL retains all sample types)}

\item{exclude_flags}{If TRUE, excludes all samples with any flag, otherwise returns all samples}
}
\description{
Loads water quality data with associated tables
}

