% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timeseries.R
\name{water_year}
\alias{water_year}
\title{Get water year from date}
\usage{
water_year(x)
}
\arguments{
\item{x}{vector of POSIXct values}
}
\value{
numeric vector of water years
}
\description{
Given vector of POSIXct values, returns vector of water years
}
\examples{
water_year(c(ymd("2000-09-30","2000-10-01","2000-10-02"))) # c(2000,2000,2001)
}

